refactoring: export to json from init.py
